Exchange a clean approach for a barbaric one
V8 has proven once again that attempting to micro-optimise based on your
assumptions will inevitably result in you looking like an idiot.

First of all, the "clean" state-driven approach that was being attempted
generated a catastrophic mess when run on ee2946's diff. Sensing there'd
be hours of effort with getting this approach working correctly, I tried
a more brutal approach. Which is what we have before us.

Benchmarking both approaches against a loop of 99999 cycles shows meagre
improvement - the new code is only 97.85% the speed of the first, at the
expense of 100% accuracy and clearer source.

Moral of the story: Believe in V8, screw what you think you know (unless
you're 100% sure you're right. Which in this case, I wasn't). The actual
stats of each benchmark are as follows, for posterity's sake:

┌─────────┬────────────────────────────────────────────────────────────┐
│ Before: │ 9139, 8987, 9111, 9194, 8996, 9111, 9365, 9223, 9153, 9077 │
├─────────┼────────────────────────────────────────────────────────────┤
│ After:  │ 9166, 9602, 9198, 9251, 9094, 9845, 9397, 9424, 9210, 9176 │
└─────────┴────────────────────────────────────────────────────────────┘

Times depicted in milliseconds.

Statistics depicted because I have a dumb obsession with lining up text,
and drawing pretty borders an' stuff.